multi-agent/pyproject.toml
[project]
name = "helpdesk-multi-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.9",
    "openai>=1.30",
    "anthropic>=0.49",
    "pydantic>=2.7",
    "tiktoken>=0.7",
    "python-dotenv>=1.0",
    "opentelemetry-sdk>=1.41.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.41.0",
    "flint-ai",
]

[project.scripts]
helpdesk-multi-agent = "src.main:main"

[tool.uv]
package = true

[tool.uv.sources]
flint-ai = { path = "/Users/shama/projects/flint-ai", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
